requirements.txt
transformers
huggingface_hub
streamlit
numpy
tensorflow
pandas
matplotlib
PyPDF2
python-docx
